Refactor: Unpack ResolvableFormula
